requirements.txt
scrapy
ipython
scrapy-playwright
python-dotenv
google-cloud-bigquery
scrapy-user-agents
scrapy-rotating-proxies
requests
scrapy-playwright
scrapy-user-agents